backend/requirements.txt
flask==3.0.0
python-dotenv==1.0.1
typing-extensions>=4.11.0,<5.0.0
pydantic>=2.7.4,<3.0.0
openai==1.64.0
langchain-core>=0.3.39
langchain-community>=0.0.39
langchain==0.3.19
langchain-openai==0.3.7
requests==2.31.0
sqlalchemy==2.0.27
sendgrid==6.11.0
jinja2==3.1.3
python-dateutil==2.8.2
flask-jwt-extended==4.6.0
bcrypt==4.1.2
stripe==8.4.0
werkzeug==3.0.1
flask-cors==4.0.0
langchain-community==0.3.18
langchain-openai==0.3.7
werkzeug==3.0.1
google-api-python-client==2.120.0

